requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
apscheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
deprecated==1.2.18
flask==3.1.1
frozenlist==1.7.0
future==1.0.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
line-bot-sdk==3.18.0
markupsafe==3.0.2
multidict==6.6.3
propcache==0.3.2
pydantic==2.11.7
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
